add: interface for fetch and separating
